+++ h2integrate/core/supported_models.py
@@ -156,6 +156,7 @@ supported_models = _ModelRegistry(
         "GenericSplitterPerformanceModel": "transporters:GenericSplitterPerformanceModel",
         "GenericTransporterPerformanceModel": "transporters:GenericTransporterPerformanceModel",
         "LinearDistanceCostModel": "transporters:LinearDistanceCostModel",
+        "LinearMassTransportCostModel": "transporters:LinearMassTransportCostModel",
         "IronTransportPerformanceComponent": "converters.iron:IronTransportPerformanceComponent",
         "IronTransportCostComponent": "converters.iron:IronTransportCostComponent",
         # Simple Summers